requirements.txt
pyjwt
pycryptodomex
requests
rich
tk
websocket-client